feat(pypi): split settings file
This commit splits the settings file. `main.py` contains settings for
the django project used by caluma service. `module.py` contains the
caluma specific settings and gets imported by `main.py`. This allows
third party projects that integrate Caluma to just import the settings
in `module.py`.
